T4 SerialX - Move each SerialX object to own source file
Hopefully will be able to get this such that only those objects that are used will be brought in by the system...
